requirements.txt
aiohttp==3.11.11
aiogram==3.17.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
PyYAML==6.0.2
python-dotenv==1.0.1

pypdf==5.1.0
langdetect==1.0.9
PyMuPDF==1.24.10
mcp>=1.2.4
google-genai>=1.4
jsonschema>=4.23.0
